+++ library/Datatypes/RadioButtonList/Editor.php
@@ -64,6 +64,7 @@ class Editor extends AbstractEditor
 
         $element->setLabel($this->getName());
         $element->setAttribute('label', $this->getProperty()->getName());
+        $element->setAttribute('description', $this->getProperty()->getDescription());
         $element->setValue($this->getValue());
 
         return $element;