+++ library/Datatypes/Textrich/Editor.php
@@ -76,6 +76,7 @@ class Editor extends AbstractEditor
         $id       = 'textrich' . $this->getProperty()->getId();
         $textrich = new Element\Textarea($this->getName());
         $textrich->setAttribute('label', $this->getProperty()->getName());
+        $textrich->setAttribute('description', $this->getProperty()->getDescription());
         $textrich->setAttribute('id', $id);
         $textrich->setAttribute('class', $id);
         $textrich->setValue($this->getProperty()->getValue());